Remove broken autoload.php require and dependency on GuzzleHttp
autoload.php was required but not included which was... frustrating.

The Client class had required an outside library, "GuzzleHttp".  Instead now we just rely on php's
curl extension which is much more likely to already be setup on an average users system.

In addition there is a working version of the test.php script which will be cleaned up and explained
in the next push.