add readme, example_config, histogram metric type, specific tests, and pushMetrics function
Added a detailed readme that covers the concept of the push gateway and why its needed here.
Included an example_config for a prometheus server to pull from the push gateway.
included @timeZynks Histogram.php and relevant code into the fork.
Split the test script into a test for each metric type to offer more guidance.
Added the pushMetrics() function which returns the ability for the client to push the serialized data to
a push gateway.